+++ app/libraries/Jentil/Setups/Customizer/AbstractControl.php
@@ -3,51 +3,8 @@ declare (strict_types = 1);
 
 namespace GrottoPress\Jentil\Setups\Customizer;
 
-use GrottoPress\Jentil\IdentityTrait;
-use WP_Customize_Manager as WPCustomizer;
+use GrottoPress\WordPress\SUV\Setups\Customizer\AbstractControl as Control;
 
-abstract class AbstractControl
+abstract class AbstractControl extends Control
 {
-    use IdentityTrait;
-
-    /**
-     * @var AbstractCustomizer
-     */
-    protected $customizer;
-
-    /**
-     * @var \WP_Customize_Control
-     */
-    protected $object;
-
-    /**
-     * @var mixed[string]
-     */
-    protected $args = [];
-
-    public function __construct(AbstractCustomizer $customizer)
-    {
-        $this->customizer = $customizer;
-    }
-
-    /**
-     * Get control, if already added
-     */
-    public function get(WPCustomizer $wp_customizer)
-    {
-        return $wp_customizer->get_control($this->id);
-    }
-
-    public function add(WPCustomizer $wp_customizer)
-    {
-        $wp_customizer->add_control(($this->object ?: $this->id), $this->args);
-    }
-
-    /**
-     * Remove control, if already added
-     */
-    public function remove(WPCustomizer $wp_customizer)
-    {
-        $wp_customizer->remove_control($this->id);
-    }
 }